Remove `system-clipboard` from the default build
This feature tries to connect reedline with the system clipboard for
three special bindings.
To do so it uses the `arboard` crate with heavy dependencies for the
system X or Wayland server or the Windows APIs. We had issues in the
headless CI with it and builds with musl seem to stall.

Removing it from the default build should negatively impact only a small
subset of users aware of the extra bindings. You can still use the
internal clipboard for binding based selection and the terminals extra
bindings to copy arbitrary content into the system clipboard.

For all other users it removes potential sources of failure and a whole
1 MB of release mode binary size. Furthermore a potentially substantial
attack surface for Nushell is gone for default builds.

If you strictly depend on this behavior, you can still build with the
addition of `--features system-clipboard`