Fix handling of stopped TUI applications on unix
Instead of handling a foreground process being stopped in any way, we were
simply ignoring SIGTSTP (which the terminal usually sends to the foreground
process group when Ctrl-Z is pressed), and propagating this to our foreground
children. This works for most processes, but it generally fails for
applications which put the terminal in raw mode[1] and implement their own
suspension mechanism (typically TUI apps like helix[2], neovim[3], top[4] or
amp[5]). In these cases, triggering suspension within the app results in the
terminal getting blocked, since the application is waiting for a SIGCONT, while
nushell is waiting for it to exit.

Fix this by unblocking SIGTSTP for our foreground children (neovim, helix and
probably others send this to themselves while trying to suspend), and
displaying the following message whenever one of them gets stopped:

    nushell currently does not support background jobs
    press any key to continue

Pressing any key will then send SIGCONT to the child's process group and resume
waiting.

This fix is probably going to be superseded by a proper background job
implementation (#247) at some point, but for now it's better than completely
blocking the terminal.

[1] https://docs.rs/crossterm/latest/crossterm/terminal/index.html#raw-mode
[2] https://helix-editor.com/
[3] https://neovim.io/
[4] https://man7.org/linux/man-pages/man1/top.1.html
[5] https://amp.rs/

fixes #1038
fixes #7335
fixes #10335